+++ Application/Commands/SingleOperand/SingleOperandCommand.cs
@@ -4,7 +4,7 @@ namespace Application.Commands.SingleOperand
 {
     public class SingleOperandCommand : IRequest<long>
     {
-        public int Value { get; set; }
+        public long Value { get; set; }
         public string Operator { get; set; }
     }
 }
