+++ Application/Interfaces/ICalculationService.cs
@@ -2,11 +2,11 @@
 {
     public interface ICalculationService
     {
-        float Add(float firstAddend, float secondAddend);
-        float Subtract(float minuend, float subtrahend);
-        float Multiply(float multiplier, float multiplicand);
-        float Divide(float dividend, float divisor);
-        long Factorial(int value);
-        long GetFibonacciAt(int index);
+        decimal Add(decimal firstAddend, decimal secondAddend);
+        decimal Subtract(decimal minuend, decimal subtrahend);
+        decimal Multiply(decimal multiplier, decimal multiplicand);
+        decimal Divide(decimal dividend, decimal divisor);
+        long Factorial(long value);
+        long GetFibonacciAt(long index);
     }
 }
